// src/types/research.ts
export type Research = {
    id: number
    slug: string
    title: string
    description: string
    author: string
    otherNames: string[]
    printHouse: string
    printDate: string
    language: string
    pages: number
    parts: number
    views: number
    image: string
    pdf: string
}